WorkReportAgent/etc/tmp/vision_test/requirements.txt
google-genai
pyyaml
opencv-python

